+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Fixed
 - Error message WHERE token = 'xxx' (#595)
+- Extend length if log IP address storage (#606)
 - Make synchronize LS surveys multibyte save (#598)
 - The default roles no longer show the ask screen (#604)